refactor: 폴더명 수정 (contexts -> constants)
