+++ components/src/lib/index.ts
@@ -1,2 +1,2 @@
 export { default as useModal } from './hooks/useModal';
-export { default as ModalContext } from './contexts/modalContext';
+export { default as ModalContext } from './constants/modalContext';
