+++ hooks/src/lib/constants/index.ts
@@ -1,3 +1,3 @@
 export * from './card';
-export * from './constants';
 export * from './regExp';
+export * from './system';
