+++ hooks/src/lib/constants/system.ts
@@ -1,0 +1,2 @@
+export const INVALID_INPUT_VALUE = '';
+export const CARD_HOLDER_MAX_LENGTH = 50;
