--- hooks/src/lib/contexts/constants.ts
@@ -1,5 +1,0 @@
-export const VALID_LENGTH = {
-  cardNumber: 4,
-  cvc: 3,
-  password: 2,
-};
